reactor component MTE implementation

+++ src/main/java/gregtech/api/fission/component/NeutronEmitter.java
@@ -5,5 +5,5 @@ public interface NeutronEmitter extends FissionComponent {
     /**
      * @return the generated amount of neutrons
      */
-    int generateNeutrons();
+    float generateNeutrons();
 }
